+++ src/api/apiHelpers.native.test.ts
@@ -1,4 +1,3 @@
-import AsyncStorage from '@react-native-async-storage/async-storage'
 import { Platform } from 'react-native'
 import CodePush from 'react-native-code-push'
 
@@ -12,15 +11,14 @@ import { mockServer } from 'tests/mswServer'
 
 import { ApiError } from './ApiError'
 import {
-  removeRefreshedAccessToken,
   createNeedsAuthenticationResponse,
   handleGeneratedApiResponse,
   isAPIExceptionCapturedAsInfo,
-  refreshAccessToken,
   RefreshTokenExpiredResponse,
   safeFetch,
 } from './apiHelpers'
 import { Configuration, DefaultApi, RefreshResponse } from './gen'
+import { removeRefreshedAccessToken } from './refreshAccessToken'
 
 jest.spyOn(PackageJson, 'getAppVersion').mockReturnValue('1.10.5')
 
@@ -66,7 +64,6 @@ describe('[api] helpers', () => {
   const mockFetch = jest.spyOn(global, 'fetch')
   const mockGetTokenStatus = jest.spyOn(jwt, 'getTokenStatus')
   const mockGetRefreshToken = jest.spyOn(Keychain, 'getRefreshToken')
-  const mockClearRefreshToken = jest.spyOn(Keychain, 'clearRefreshToken')
 
   afterEach(() => {
     mockFetch.mockReset()
@@ -328,109 +325,6 @@ describe('[api] helpers', () => {
         },
       })
       expect(response).toEqual(await expectedResponse)
-    })
-  })
-
-  describe('refreshAccessToken', () => {
-    it('should remove access token when there is no refresh token', async () => {
-      mockGetRefreshToken.mockResolvedValueOnce(null)
-
-      await refreshAccessToken(api, 0)
-
-      expect(AsyncStorage.removeItem).toHaveBeenCalledWith('access_token')
-    })
-
-    it('should return FAILED_TO_GET_REFRESH_TOKEN_ERROR when there is no refresh token', async () => {
-      mockGetRefreshToken.mockResolvedValueOnce(null)
-
-      const result = await refreshAccessToken(api, 0)
-
-      expect(result).toEqual({ error: 'Erreur lors de la récupération du refresh token' })
-    })
-
-    it('should store the new access token when everything is ok', async () => {
-      const password = 'refreshToken'
-      mockGetRefreshToken.mockResolvedValueOnce(password)
-      mockFetch.mockResolvedValueOnce(respondWith({ accessToken }))
-
-      await refreshAccessToken(api, 0)
-
-      expect(AsyncStorage.setItem).toHaveBeenCalledWith('access_token', accessToken)
-    })
-
-    it('should return the new access token when everything is ok', async () => {
-      const password = 'refreshToken'
-      mockGetRefreshToken.mockResolvedValueOnce(password)
-      mockFetch.mockResolvedValueOnce(respondWith({ accessToken }))
-
-      const result = await refreshAccessToken(api, 0)
-
-      expect(result).toEqual({ result: accessToken })
-    })
-
-    it('should remove tokens when there is an unexpected behavior', async () => {
-      const password = 'refreshToken'
-      mockGetRefreshToken.mockResolvedValueOnce(password)
-      mockFetch.mockResolvedValueOnce(respondWith({ error: 'server error' }, 500))
-
-      await refreshAccessToken(api, 0)
-
-      expect(AsyncStorage.removeItem).toHaveBeenCalledWith('access_token')
-      expect(mockClearRefreshToken).toHaveBeenCalledTimes(1)
-    })
-
-    it('should return UNKNOWN_ERROR_WHILE_REFRESHING_ACCESS_TOKEN when there is an unexpected behavior', async () => {
-      const password = 'refreshToken'
-      mockGetRefreshToken.mockResolvedValueOnce(password)
-      mockFetch.mockResolvedValueOnce(respondWith({ error: 'server error' }, 500))
-
-      const result = await refreshAccessToken(api, 0)
-
-      expect(result).toStrictEqual({
-        error: 'Une erreur inconnue est survenue lors de la regénération de l’access token',
-      })
-    })
-
-    it('should retry to refresh access token when the first call fails', async () => {
-      const password = 'refreshToken'
-      mockGetRefreshToken.mockResolvedValueOnce(password)
-      mockFetch
-        .mockResolvedValueOnce(respondWith({ error: 'server error' }, 500))
-        .mockResolvedValueOnce(respondWith({ accessToken }))
-
-      const result = await refreshAccessToken(api, 1)
-
-      expect(result).toEqual({ result: accessToken })
-    })
-
-    it('should remove tokens when refresh token is expired', async () => {
-      const password = 'refreshToken'
-      mockGetRefreshToken.mockResolvedValueOnce(password)
-      mockFetch.mockResolvedValueOnce(respondWith({ error: 'refresh token is expired' }, 401))
-
-      await refreshAccessToken(api, 0)
-
-      expect(AsyncStorage.removeItem).toHaveBeenCalledWith('access_token')
-      expect(mockClearRefreshToken).toHaveBeenCalledTimes(1)
-    })
-
-    it('should return REFRESH_TOKEN_IS_EXPIRED_ERROR when refresh token is expired', async () => {
-      const password = 'refreshToken'
-      mockGetRefreshToken.mockResolvedValueOnce(password)
-      mockFetch.mockResolvedValueOnce(respondWith({ error: 'refresh token is expired' }, 401))
-
-      const result = await refreshAccessToken(api, 0)
-
-      expect(result).toEqual({ error: 'Le refresh token est expiré' })
-    })
-
-    it("should log to sentry when can't refresh token", async () => {
-      const error = new Error('Error')
-      mockFetch.mockRejectedValueOnce(error)
-
-      await refreshAccessToken(api, 0)
-
-      expect(eventMonitoring.captureException).toHaveBeenCalledWith(error)
     })
   })
